Renamed capability file name.
