--- src/NTwain/Triplets/DGCustom.cs
@@ -1,16 +1,0 @@
-﻿//using System;
-//using System.Collections.Generic;
-//using System.Linq;
-//using System.Text;
-
-//namespace NTwain.Triplets
-//{
-//    /// <summary>
-//    /// Provides direct access to the triplet call.
-//    /// </summary>
-//    public partial class DGCustom : BaseTriplet
-//    {
-//        internal DGCustom(TwainSession session) : base(session) { }
-
-//    }
-//}
